refactor order

+++ src/main/java/com/macoredroid/onlinebookstore/info/bookinfo.java
@@ -4,11 +4,11 @@ public class bookinfo {
     public int booklistID;
     public String name;
     public String author;
-    public Double price;
+    public int price;
     public String isbn;
     public int stock;
     public int sales;
-    public bookinfo(int booklistID,String name,String author,Double price,String isbn,int stock,int sales)
+    public bookinfo(int booklistID,String name,String author,int price,String isbn,int stock,int sales)
     {
         this.booklistID=booklistID;
         this.name=name;
